suspend: make sync() on suspend-to-RAM optional
Linux suspend-to-RAM was unreliable when first developed,
and so sys_sync() was invoked inside the kernel at the
start of every suspend flow.

Today, many devices invoke suspend with high reliability
and high frequency. They may not want to be forced to pay
the performance cost of sync() on every suspend.

So here we make the file system sync() optional. Just
de-select CONFIG_SUSPEND_FS_SYNC to delete the call. Keep
CONFIG_SUSPEND_FS_SYNC to keep the call, and also includes
the new sysfs attribte which can be cleared to disable the
file-system sync:

/sys/power/suspend_fs_sync

Sync disabled by default, but config is enabled.

Change-Id: I93f08bf6f6113cfa70fb3458e9199a67e7ccebef
Signed-off-by: Len Brown <len.brown@intel.com>
Signed-off-by: mydongistiny <jaysonedson@gmail.com>