sched/rt: fix schedtune accouting on prio change
When we change the RT priority of a task which is currently running, the
task is dequeued from the old priority queue to be enqueued again in the
new one. However, the current enqueue_task_rt disregards completely
tasks which are currently RUNNING.

This introduces a double dequeue for RT tasks, which can be observed in
the current kernel for example when a Binder worker has its priority
changed right before suspending while waiting for more work.

The double decrement is guarded by a min capping to zero, however the
SchedTune functional behaviors are compromised because this behavior can
disable a boostgroup while there are still some tasks RUNNABLE in the
same group.

More in general the asymmetry between {en,de}queue_task_rt() seems to
not make sense since these two methods are always called in pairs for
tasks entering/exiting a CPU for different reasons (wakeup/suspend,
migration, CGroup change).

This patch fixes the SchedTune accounting bug by removing this asymmetry
and unconditionally updating the counter at each enqueue/dequeue call
for RT tasks, which is exactly what we already do for FAIR tasks too.

Bug: 30210506

Change-Id: I029c5bcb41e955f42472abe2a55bb080f1b043ba
Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>
Signed-off-by: mydongistiny <jaysonedson@gmail.com>